bot/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
python-dotenv==1.0.0

# aioredis==1.3.1
# SQLAlchemy
# asyncpg  # для работы Алхимии

# selenium==4.0.0a7
# beautifulsoup4
# requests==2.25.1
# emoji

# FIXME: Не получается обновить pip, какая-то ошибка ENVS